Moved logic of HTML generation of suite result to visualization
